+++ build.gradle
@@ -12,7 +12,7 @@ version = '1.0.0'
 sourceCompatibility = '11'
 targetCompatibility = '11'
 
-mainClassName = 'uk.gov.hmcts.reform.fpl.ccddatamigration.CcdDataMigrationApplication'
+mainClassName = 'uk.gov.hmcts.reform.fpl.ccddatamigration.DataMigrationProcessor'
 
 compileJava {
     options.compilerArgs << '-parameters' << '-Xlint:deprecation'
