+++ src/main/java/uk.gov.hmcts.reform.fpl.ccddatamigration/DataMigrationProcessor.java
@@ -1,20 +1,18 @@
-package uk.gov.hmcts.reform.fpl.ccddatamigration.processor;
+package uk.gov.hmcts.reform.fpl.ccddatamigration;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.boot.SpringApplication;
-import org.springframework.context.annotation.Configuration;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.PropertySource;
-import org.springframework.stereotype.Component;
 import uk.gov.hmcts.reform.authorisation.generators.AuthTokenGenerator;
 import uk.gov.hmcts.reform.fpl.ccddatamigration.service.MigrationService;
 import uk.gov.hmcts.reform.idam.client.IdamClient;
 
 @Slf4j
-@Component
-@Configuration
+@SpringBootApplication
 @PropertySource("classpath:application.properties")
 public class DataMigrationProcessor implements CommandLineRunner {
 
@@ -24,8 +22,8 @@ public class DataMigrationProcessor implements CommandLineRunner {
     @Value("${idam.password}")
     private String idamPassword;
 
-    @Value("${ccd.jurisdictionid}")
-    private String jurisdictionId;
+    @Value("${ccd.jurisdiction}")
+    private String jurisdiction;
 
     @Value("${ccd.casetype}")
     private String caseType;
@@ -49,51 +47,50 @@ public class DataMigrationProcessor implements CommandLineRunner {
     private MigrationService migrationService;
 
     public static void main(String[] args) {
+        System.setProperty("http.proxyHost", "proxyout.reform.hmcts.net");
+        System.setProperty("http.proxyPort", "8080");
+        System.setProperty("https.proxyHost", "proxyout.reform.hmcts.net");
+        System.setProperty("https.proxyPort", "8080");
+
         SpringApplication.run(DataMigrationProcessor.class, args);
     }
 
     @Override
     public void run(String... args) {
-
         try {
             if (debugEnabled) {
-                log.info("Start processing cases");
+                log.info("Data migration started");
             }
             String userToken = idamClient.authenticateUser(idamUsername, idamPassword);
             if (debugEnabled) {
-                log.info("  userToken  : {}", userToken);
+                log.info("  userToken: {}", userToken);
             }
             String s2sToken = authTokenGenerator.generate();
             if (debugEnabled) {
-                log.info("  s2sToken : {}", s2sToken);
+                log.info("  s2sToken: {}", s2sToken);
             }
             String userId = idamClient.getUserDetails(userToken).getId();
             if (debugEnabled) {
-                log.info("  userId  : {}", userId);
+                log.info("  userId: {}", userId);
             }
 
             if (ccdCaseId != null && !ccdCaseId.isBlank()) {
-                log.info("migrate case, caseId  {}", ccdCaseId);
                 migrationService.processSingleCase(userToken, s2sToken, ccdCaseId);
             } else {
-                migrationService.processAllTheCases(userToken, s2sToken, userId, jurisdictionId, caseType);
+                migrationService.processAllTheCases(userToken, s2sToken, userId, jurisdiction, caseType);
             }
-            log.info("Migrated Cases {} ",
-                migrationService.getMigratedCases() != null && !migrationService.getMigratedCases().isBlank() ? migrationService.getMigratedCases() : "NONE");
+
+            log.info("Migrated cases: {} ", !migrationService.getMigratedCases().isEmpty() ? migrationService.getMigratedCases() : "NONE");
+            log.info("Failed cases: {}", !migrationService.getFailedCases().isEmpty() ? migrationService.getFailedCases() : "NONE");
 
             log.info("-----------------------------");
             log.info("Data migration completed");
             log.info("-----------------------------");
-            log.info("Total number of cases: " + migrationService.getTotalNumberOfCases());
-            log.info("Total migrations performed: " + migrationService.getTotalMigrationsPerformed());
+            log.info("Total number of cases: {}", migrationService.getMigratedCases().size() + migrationService.getFailedCases().size());
+            log.info("Total migrations performed: {}", migrationService.getMigratedCases().size());
             log.info("-----------------------------");
-            log.info("Failed Cases {}",
-                migrationService.getFailedCases() != null && !migrationService.getFailedCases().isBlank() ? migrationService.getFailedCases() : "NONE");
-
         } catch (Throwable e) {
-            log.error("Migration failed with the following reason :", e.getMessage());
-            e.printStackTrace();
+            log.error("Migration failed with the following reason: {}", e.getMessage(), e);
         }
     }
-
 }
