+++ src/main/java/uk.gov.hmcts.reform.fpl.ccddatamigration/ccd/CoreCaseDataService.java
@@ -1,25 +1,42 @@
 package uk.gov.hmcts.reform.fpl.ccddatamigration.ccd;
 
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
+import uk.gov.hmcts.reform.authorisation.generators.AuthTokenGenerator;
+import uk.gov.hmcts.reform.ccd.client.CoreCaseDataApi;
 import uk.gov.hmcts.reform.ccd.client.model.CaseDataContent;
 import uk.gov.hmcts.reform.ccd.client.model.CaseDetails;
 import uk.gov.hmcts.reform.ccd.client.model.Event;
 import uk.gov.hmcts.reform.ccd.client.model.StartEventResponse;
+import uk.gov.hmcts.reform.idam.client.IdamClient;
 import uk.gov.hmcts.reform.idam.client.models.UserDetails;
 
-@Service
-public class CcdUpdateServiceImpl extends BaseCcdCaseService implements CcdUpdateService {
+import static uk.gov.hmcts.reform.fpl.ccddatamigration.s2s.AuthUtil.getBearerToken;
 
-    @Override
-    public CaseDetails update(String caseId, Object data, String eventId, String authorisation,
-                              String eventSummary, String eventDescription) {
-        UserDetails userDetails = getUserDetails(authorisation);
+@Service
+public class CoreCaseDataService {
+
+    @Value("${ccd.jurisdiction}")
+    private String jurisdiction;
+    @Value("${ccd.casetype}")
+    private String caseType;
+
+    @Autowired
+    private IdamClient idamClient;
+    @Autowired
+    private AuthTokenGenerator authTokenGenerator;
+    @Autowired
+    private CoreCaseDataApi coreCaseDataApi;
+
+    public CaseDetails update(String authorisation, String caseId, String eventId, String eventSummary, String eventDescription, Object data) {
+        UserDetails userDetails = idamClient.getUserDetails(getBearerToken(authorisation));
 
         StartEventResponse startEventResponse = coreCaseDataApi.startEventForCaseWorker(
-            getBearerUserToken(authorisation),
-            getServiceAuthToken(),
+            getBearerToken(authorisation),
+            authTokenGenerator.generate(),
             userDetails.getId(),
-            jurisdictionId,
+            jurisdiction,
             caseType,
             caseId,
             eventId);
@@ -36,10 +53,10 @@ public class CcdUpdateServiceImpl extends BaseCcdCaseService implements CcdUpdat
             .build();
 
         return coreCaseDataApi.submitEventForCaseWorker(
-            getBearerUserToken(authorisation),
-            getServiceAuthToken(),
+            getBearerToken(authorisation),
+            authTokenGenerator.generate(),
             userDetails.getId(),
-            jurisdictionId,
+            jurisdiction,
             caseType,
             caseId,
             true,
