+++ src/main/java/uk.gov.hmcts.reform.fpl.ccddatamigration/s2s/AuthUtil.java
@@ -9,7 +9,7 @@ public class AuthUtil {
 
     private static final String BEARER = "Bearer ";
 
-    public static String getBearToken(String token) {
+    public static String getBearerToken(String token) {
         if (token == null || token.isBlank()) {
             return token;
         }
