+++ src/main/resources/application.properties
@@ -13,7 +13,7 @@ idam.s2s-auth.microservice=fpl_case_service
 idam.s2s-auth.totp_secret=AABBCCDDEEFFGGHH
 idam.s2s-auth.url=http://localhost:4502
 
-ccd.jurisdictionid=PUBLICLAW
+ccd.jurisdiction=PUBLICLAW
 ccd.casetype=CARE_SUPERVISION_EPO
 ccd.caseId=
 ccd.dryrun=true
